requirements.txt
aliyun-python-sdk-alidns==2.0.13
aliyun-python-sdk-core==2.13.10
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-ecs==4.10.1
aliyun-python-sdk-slb==3.2.15
amqp==2.5.1
anyjson==0.3.3
asn1crypto==0.24.0
Babel==2.7.0
bcrypt==3.1.7
billiard==3.5.0.3
celery==4.1.0
cffi==1.12.3
cryptography==2.7
Django==2.1.12
django-celery-beat==1.1.0
django-redis==4.10.0
django-timezone-field==3.0
dwebsocket==0.5.11
ephem==3.7.7.0
flower==0.9.3
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
importlib-metadata==0.23
jmespath==0.9.4
kombu==4.0.2
more-itertools==7.2.0
mysqlclient==1.3.14
paramiko==2.6.0
pycparser==2.19
pycryptodome==3.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
vine==1.3.0
zipp==0.6.0
